+++ composer.json
@@ -70,7 +70,7 @@
         "codeception/codeception": "3.1.1",
         "doctrine/doctrine-fixtures-bundle": "3.2.2",
         "friendsofphp/php-cs-fixer": "2.15.3",
-        "friendsoftwig/twigcs": "3.1.3",
+        "friendsoftwig/twigcs": "^3.1",
         "fzaninotto/faker": "1.8.0",
         "maglnet/composer-require-checker": "2.0.0",
         "pdepend/pdepend": "2.5.2",
@@ -80,7 +80,7 @@
         "phpstan/phpstan-symfony": "0.11.6",
         "povils/phpmnd": "2.1.0",
         "sebastian/phpcpd": "4.1.0",
-        "sensiolabs/security-checker": "6.0.2",
+        "sensiolabs/security-checker": "^6.0",
         "squizlabs/php_codesniffer": "3.4.2",
         "symfony/browser-kit": "4.3.4",
         "symfony/css-selector": "4.3.4",
