+++ src/Entity/Traits/Bookmark.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Labstag\Entity\Traits;
+
+use Doctrine\Common\Collections\Collection;
+use Labstag\Entity\Bookmark as BookmarkEntity;
+
+trait Bookmark
+{
+    /**
+     * @return BookmarkEntity|Collection
+     */
+    public function getBookmarks(): Collection
+    {
+        return $this->bookmarks;
+    }
+
+    public function addBookmark(BookmarkEntity $bookmark): self
+    {
+        if (!$this->bookmarks->contains($bookmark)) {
+            $this->bookmarks[] = $bookmark;
+        }
+
+        return $this;
+    }
+
+    public function removeBookmark(BookmarkEntity $bookmark): self
+    {
+        if ($this->bookmarks->contains($bookmark)) {
+            $this->bookmarks->removeElement($bookmark);
+        }
+
+        return $this;
+    }
+}
